Convert to using jinja2 templates to generate Dockerfiles
